pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hackathon-sakhi"
version = "0.2.2"
description = "Women Safety MCP Servers - Hackathon Project (Weather, News Dashboard, Telegram Alerts, Location Monitor)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Hackathon Team" }
]
keywords = ["mcp", "women-safety", "hackathon", "telegram", "weather", "news"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "feedparser>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.urls]
Homepage = "https://pypi.org/project/hackathon-sakhi/"
Repository = "https://pypi.org/project/hackathon-sakhi/"

[project.scripts]
# Entry points for each MCP server
sakhi-weather = "hackathon_sakhi.weather:main"
sakhi-news = "hackathon_sakhi.news:main"
sakhi-telegram = "hackathon_sakhi.telegram:main"
sakhi-location = "hackathon_sakhi.location_v2:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hackathon_sakhi"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]
